refactor: CourseFilterAction.UpdateLengthRange 처리 시 불필요한 타입 변환 제거

+++ android/app/src/main/java/io/coursepick/coursepick/presentation/course/CoursesViewModel.kt
@@ -425,7 +425,7 @@ class CoursesViewModel
                 }
 
                 is CourseFilterAction.UpdateLengthRange -> {
-                    updateLengthRange(action.start.toDouble(), action.end.toDouble())
+                    updateLengthRange(action.start, action.end)
                 }
 
                 is CourseFilterAction.ToggleDifficulty -> {
